% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcourse.R
\docType{package}
\name{pkgcourse}
\alias{pkgcourse}
\alias{pkgcourse-package}
\title{pkgcourse: A package created on Mango Solutions packgaes course.}
\description{
Tools for doing basic analysis.
}
\author{
Tony Carter \email{tdcarter1985@gmail.com}
}
