% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsummary.R
\name{statSummary}
\alias{statSummary}
\title{statSummary}
\usage{
statSummary(x, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{numeric} vector for summarising}

\item{na.rm}{a \code{logical} value stating whether to remove \code{NA}
values; defaults to \code{TRUE} - remove \code{NA}s}
}
\value{
Returns a vector of summary statistics for \code{x}:
\describe{
\item{Min}{Minimum value of data}
\item{Mean}{Mean of the data}
}
}
\description{
statSummary
}
\author{
Tony Carter <tdcarter1985@gmail.com>
}
